+++ src/error.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 export abstract class AiScriptError extends Error {
 	// name is read by Error.prototype.toString
 	public name = 'AiScript';
@@ -57,6 +59,15 @@ export class AiScriptRuntimeError extends AiScriptError {
  * RuntimeError for illegal access to arrays.
  */
 export class AiScriptIndexOutOfRangeError extends AiScriptRuntimeError {
+	constructor(message: string, info?: any) {
+		super(message, info);
+	}
+}
+/**
+ * Errors thrown by users.
+ */
+export class AiScriptUserError extends AiScriptRuntimeError {
+	public name = '';
 	constructor(message: string, info?: any) {
 		super(message, info);
 	}